Tool-Based Agent/requirements.txt
duckduckgo-search
langchain
langchain-core
langchain-groq
python-dotenv
requests